+++ MANIFEST.in
@@ -67,8 +67,8 @@ include psutil/arch/openbsd/disk.c
 include psutil/arch/openbsd/disk.h
 include psutil/arch/openbsd/mem.c
 include psutil/arch/openbsd/mem.h
-include psutil/arch/openbsd/specific.c
-include psutil/arch/openbsd/specific.h
+include psutil/arch/openbsd/proc.c
+include psutil/arch/openbsd/proc.h
 include psutil/arch/osx/cpu.c
 include psutil/arch/osx/cpu.h
 include psutil/arch/osx/process_info.c
