+++ setup.py
@@ -223,7 +223,7 @@ elif OPENBSD:
             'psutil/arch/openbsd/cpu.c',
             'psutil/arch/openbsd/disk.c',
             'psutil/arch/openbsd/mem.c',
-            'psutil/arch/openbsd/specific.c',
+            'psutil/arch/openbsd/proc.c',
         ],
         define_macros=macros,
         libraries=["kvm"])
